+++ src/main/java/ortus/boxlang/lsp/workspace/visitors/UnusedVariableDiagnosticVisitor.java
@@ -71,11 +71,7 @@ public class UnusedVariableDiagnosticVisitor extends SourceCodeVisitor {
 	}
 
 	public void visit( BoxProperty node ) {
-		String propertyName = BLASTTools.getPropertyName( node );
-		if ( propertyName == null ) {
-			return;
-		}
-		properties.add( propertyName.toLowerCase() );
+		BLASTTools.getPropertyName( node ).ifPresent( propertyName -> properties.add( propertyName.toLowerCase() ) );
 	}
 
 	private String getNameFromNode( BoxNode node ) {